Add setback column to tidyparcel object

`add_setbacks()` returns the tidyparcel object with new columns describing the setback and units of each side.

Usage

add_setbacks(
  tidyparcel,
  tidydistrict,
  tidybuilding,
  tidyparcel_with_dims = NULL,
  zoning_req = NULL
)

Arguments

tidyparcel

A tidyparcel object is an simple features object depicting each side of a parcel and its label (front, Interior side, Exterior side, rear, centroid).

tidydistrict

The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.

tidybuilding

A tidybuilding is a list of data frames used to represent a building.

Value

Returns the tidyparcel object with a "setbacks" and "units" column added to the end.