Compare building land use and allowed land uses

`check_land_use()` takes a tidybuilding and a tidydistrict to see if the district's zoning code allows the tidybuilding based on land use.

Usage

check_land_use(
  tidybuilding,
  tidydistrict,
  tidyparcel = NULL,
  zoning_req = NULL
)

Arguments

tidybuilding

A tidybuilding is a list of data frames used to represent a building.

tidydistrict

The tidydistrict corresponding to the tidyparcel. A tidydistrict object is one row from a tidyzoning simple features object.

tidyparcel

A tidyparcel object is an simple features object depicting each side of a parcel and its label (front, Interior side, Exterior side, rear, centroid).

Value

Returns TRUE or FALSE stating whether or not the building would be allowed in the district based on land use. Note: If there is no recorded land use requirement in zoning code, it returns FALSE